One way to get the icon where I want in the actionbar but prob not what I'm gonna end up doing
The asset provided by coin is too big and causes the actionbar to have
extra space in the titlebar. Normally I would just ask for a different
asset but the request is to use the provided assets so I'll have to get
creative. This is one way - custom toolbar.
I'm gonna try to resize the asset in code before giving it to the
actionbar next.